second attempt for excel-download
