+++ fi_db.json
@@ -1,1 +1,1 @@
-{"_default": {"1": {"advisory_fee": "20.0", "chisai_kinri": "0.015000", "chisai_shoukan_kikan": 20, "chisai_sueoki_years": 2, "const_start_date_year": 2026, "const_start_date_month": 8, "const_start_date_day": 8, "const_start_date": "2026-08-08", "const_years": 2, "discount_rate": "0.033836", "first_end_fy": "2027-03-31", "fudousanshutokuzei_hyoujun": "0.0", "fudousanshutokuzei_ritsu": "0.0", "growth": "0.000000", "hojo_ritsu": "0.300000", "houjinzei_ritsu": "0.0", "houjinjuminzei_kintou": "0.0", "houjinjuminzei_ritsu_todouhuken": "0.0", "houjinjuminzei_ritsu_shikuchoson": "0.0", "ijikanri_unnei": "70.000000", "ijikanri_unnei_LCC": "66.800000", "ijikanri_unnei_org": "70.000000", "ijikanri_unnei_org_LCC": "66.800000", "ijikanri_unnei_1": "40.000000", "ijikanri_unnei_1_LCC": "38.000000", "ijikanri_unnei_1_org": "40.000000", "ijikanri_unnei_1_org_LCC": "38.000000", "ijikanri_unnei_2": "20.000000", "ijikanri_unnei_2_LCC": "19.000000", "ijikanri_unnei_2_org": "20.000000", "ijikanri_unnei_2_org_LCC": "19.000000", "ijikanri_unnei_3": "10.000000", "ijikanri_unnei_3_LCC": "9.800000", "ijikanri_unnei_3_org": "10.000000", "ijikanri_unnei_3_org_LCC": "9.800000", "ijikanri_unnei_years": 16, "kappu_kinri_spread": "0.000000", "Kappu_kinri": "0.026950", "kijun_kinri": "0.01695038", "kisai_jutou": "0.750000", "kisai_koufu": "0.300000", "kitai_bukka": "0.01688591847991190331157440596", "koteishisanzei_hyoujun": "0.0", "koteishisanzei_ritsu": "0.0", "lg_spread": "0.010000", "mgmt_type": "\u5e02\u753a\u6751", "monitoring_costs_PSC": "10.0", "monitoring_costs_LCC": "6.0", "option_02": "0", "pre_kyoukouka": true, "proj_ctgry": "\u30b5\u30fc\u30d3\u30b9\u8cfc\u5165\u578b", "proj_type": "DBO(SPC\u306a\u3057)", "proj_years": 18, "rakusatsu_ritsu": "0.950000", "reduc_shisetsu": "0.050000", "reduc_ijikanri_1": "0.050000", "reduc_ijikanri_2": "0.050000", "reduc_ijikanri_3": "0.020000", "riyouryoukin_shunyu": "0.0", "shisetsu_seibi": "4000.000000", "shisetsu_seibi_LCC": "3800.000000", "shisetsu_seibi_org": "4000.000000", "shisetsu_seibi_org_LCC": "3800.000000", "shisetsu_seibi_paymentschedule_ikkatsu": "1.000000", "shisetsu_seibi_paymentschedule_kappu": "0.000000", "shoukan_kaishi_jiki": 5, "SPC_keihi": "0", "SPC_fee": "0", "SPC_shihon": "0", "SPC_yobihi": "0", "SPC_hiyou_atsukai": 1, "SPC_hiyou_total": "0", "SPC_hiyou_nen": "0", "SPC_keihi_LCC": "0.0", "target_years": 45, "tourokumenkyozei_hyoujun": "0.0", "tourokumenkyozei_ritsu": "0.0", "yosantanka_hiritsu_shisetsu": "0.000000", "yosantanka_hiritsu_ijikanri_1": "0.000000", "yosantanka_hiritsu_ijikanri_2": "0.000000", "yosantanka_hiritsu_ijikanri_3": "0.000000", "zei_total": "0.180000"}}}
+{"_default": {"1": {"advisory_fee": "20.0", "chisai_kinri": "0.015000", "chisai_shoukan_kikan": 20, "chisai_sueoki_years": 2, "const_start_date_year": 2026, "const_start_date_month": 11, "const_start_date_day": 11, "const_start_date": "2026-11-11", "const_years": 2, "discount_rate": "0.030511", "first_end_fy": "2027-03-31", "fudousanshutokuzei_hyoujun": "0.0", "fudousanshutokuzei_ritsu": "0.0", "growth": "0.000000", "hojo_ritsu": "0.300000", "houjinzei_ritsu": "0.0", "houjinjuminzei_kintou": "0.0", "houjinjuminzei_ritsu_todouhuken": "0.0", "houjinjuminzei_ritsu_shikuchoson": "0.0", "ijikanri_unnei": "53.000000", "ijikanri_unnei_LCC": "50.590000", "ijikanri_unnei_org": "53.000000", "ijikanri_unnei_org_LCC": "50.590000", "ijikanri_unnei_1": "30.000000", "ijikanri_unnei_1_LCC": "28.500000", "ijikanri_unnei_1_org": "30.000000", "ijikanri_unnei_1_org_LCC": "28.500000", "ijikanri_unnei_2": "15.000000", "ijikanri_unnei_2_LCC": "14.250000", "ijikanri_unnei_2_org": "15.000000", "ijikanri_unnei_2_org_LCC": "14.250000", "ijikanri_unnei_3": "8.000000", "ijikanri_unnei_3_LCC": "7.840000", "ijikanri_unnei_3_org": "8.000000", "ijikanri_unnei_3_org_LCC": "7.840000", "ijikanri_unnei_years": 15, "kappu_kinri_spread": "0.000000", "Kappu_kinri": "0.023625", "kijun_kinri": "0.01362491", "kisai_jutou": "0.750000", "kisai_koufu": "0.300000", "kitai_bukka": "0.01688591847991190331157440596", "koteishisanzei_hyoujun": "0.0", "koteishisanzei_ritsu": "0.0", "lg_spread": "0.010000", "mgmt_type": "\u5e02\u753a\u6751", "monitoring_costs_PSC": "10.0", "monitoring_costs_LCC": "6.0", "option_02": "0", "pre_kyoukouka": true, "proj_ctgry": "\u30b5\u30fc\u30d3\u30b9\u8cfc\u5165\u578b", "proj_type": "DBO(SPC\u306a\u3057)", "proj_years": 17, "rakusatsu_ritsu": "0.950000", "reduc_shisetsu": "0.050000", "reduc_ijikanri_1": "0.050000", "reduc_ijikanri_2": "0.050000", "reduc_ijikanri_3": "0.020000", "riyouryoukin_shunyu": "0.0", "shisetsu_seibi": "3003.000000", "shisetsu_seibi_LCC": "2852.850000", "shisetsu_seibi_org": "3003.000000", "shisetsu_seibi_org_LCC": "2852.850000", "shisetsu_seibi_paymentschedule_ikkatsu": "1.000000", "shisetsu_seibi_paymentschedule_kappu": "0.000000", "shoukan_kaishi_jiki": 5, "SPC_keihi": "0", "SPC_fee": "0", "SPC_shihon": "0", "SPC_yobihi": "0", "SPC_hiyou_atsukai": 1, "SPC_hiyou_total": "0", "SPC_hiyou_nen": "0", "SPC_keihi_LCC": "0.0", "target_years": 45, "tourokumenkyozei_hyoujun": "0.0", "tourokumenkyozei_ritsu": "0.0", "yosantanka_hiritsu_shisetsu": "0.000000", "yosantanka_hiritsu_ijikanri_1": "0.000000", "yosantanka_hiritsu_ijikanri_2": "0.000000", "yosantanka_hiritsu_ijikanri_3": "0.000000", "zei_total": "0.180000"}}}
