+++ ii_db.json
@@ -1,1 +1,1 @@
-{"_default": {"1": {"mgmt_type": "\u5e02\u753a\u6751", "proj_ctgry": "\u30b5\u30fc\u30d3\u30b9\u8cfc\u5165\u578b", "proj_type": "DBO(SPC\u306a\u3057)", "proj_years": "18", "const_years": "2", "ijikanri_unnei_years": 16, "const_start_date": "2025-02-08", "kijun_kinri": "1.695038", "chisai_kinri": "1.500000", "chisai_sueoki_kikan": 2, "chisai_shoukan_kikan": 20, "zei_modori": "0.08400000000000000521804821573823573999106884002685546875", "lg_spread": "0.010000", "zei_total": "0.180000", "riyou_ryoukin": 0.0, "growth": "0.000000", "kitai_bukka": "1.688591847991190331157440596143715083599090576171875", "shisetsu_seibi": "4000.000000", "shisetsu_seibi_org": "4000.000000", "shisetsu_seibi_org_LCC": "3800.000000", "shisetsu_seibi_LCC": "3800.000000", "ijikanri_unnei_1": "40.000000", "ijikanri_unnei_1_org": "40.000000", "ijikanri_unnei_1_org_LCC": "38.000000", "ijikanri_unnei_1_LCC": "38.000000", "ijikanri_unnei_2": "20.000000", "ijikanri_unnei_2_org": "20.000000", "ijikanri_unnei_2_org_LCC": "19.000000", "ijikanri_unnei_2_LCC": "19.000000", "ijikanri_unnei_3": "10.000000", "ijikanri_unnei_3_org": "10.000000", "ijikanri_unnei_3_org_LCC": "9.800000", "ijikanri_unnei_3_LCC": "9.800000", "yosantanka_hiritsu_shisetsu": "0.000000", "yosantanka_hiritsu_ijikanri_1": "0.000000", "yosantanka_hiritsu_ijikanri_2": "0.000000", "yosantanka_hiritsu_ijikanri_3": "0.000000", "rakusatsu_ritsu": "0.950000", "reduc_shisetsu": "0.050000", "reduc_ijikanri_1": "0.050000", "reduc_ijikanri_2": "0.050000", "reduc_ijikanri_3": "0.020000", "pre_kyoukouka": true, "kisai_jutou": "0.75", "kisai_koufu": "0.299999999999999988897769753748434595763683319091796875", "hojo_ritsu": "0.299999999999999988897769753748434595763683319091796875", "zeimae_rieki": "0.000000", "SPC_keihi": "20.000000", "SPC_fee": "0.000000", "SPC_shihon": "0.000000", "SPC_yobihi": "0.000000", "SPC_hiyou_atsukai": 1, "houjinzei_ritsu": "0.0", "houjinjuminzei_kintou": "0.0", "hudousanshutokuzei_hyoujun": "0.0", "hudousanshutokuzei_ritsu": "0.0", "koteishisanzei_hyoujun": "0.0", "koteishisanzei_ritsu": "0.0", "tourokumenkyozei_hyoujun": "0.0", "tourokumenkyozei_ritsu": "0.0", "houjinjuminzei_ritsu_todouhuken": "0.0", "houjinjuminzei_ritsu_shikuchoson": "0.0"}}}
+{"_default": {"1": {"mgmt_type": "\u5e02\u753a\u6751", "proj_ctgry": "\u30b5\u30fc\u30d3\u30b9\u8cfc\u5165\u578b", "proj_type": "DBO(SPC\u306a\u3057)", "proj_years": "17", "const_years": "2", "ijikanri_unnei_years": 15, "const_start_date": "2025-02-11", "kijun_kinri": "1.362491", "chisai_kinri": "1.500000", "chisai_sueoki_kikan": 2, "chisai_shoukan_kikan": 20, "zei_modori": "0.08400000000000000521804821573823573999106884002685546875", "lg_spread": "0.010000", "zei_total": "0.180000", "riyou_ryoukin": 0.0, "growth": "0.000000", "kitai_bukka": "1.688591847991190331157440596143715083599090576171875", "shisetsu_seibi": "3003.000000", "shisetsu_seibi_org": "3003.000000", "shisetsu_seibi_org_LCC": "2852.850000", "shisetsu_seibi_LCC": "2852.850000", "ijikanri_unnei_1": "30.000000", "ijikanri_unnei_1_org": "30.000000", "ijikanri_unnei_1_org_LCC": "28.500000", "ijikanri_unnei_1_LCC": "28.500000", "ijikanri_unnei_2": "15.000000", "ijikanri_unnei_2_org": "15.000000", "ijikanri_unnei_2_org_LCC": "14.250000", "ijikanri_unnei_2_LCC": "14.250000", "ijikanri_unnei_3": "8.000000", "ijikanri_unnei_3_org": "8.000000", "ijikanri_unnei_3_org_LCC": "7.840000", "ijikanri_unnei_3_LCC": "7.840000", "yosantanka_hiritsu_shisetsu": "0.000000", "yosantanka_hiritsu_ijikanri_1": "0.000000", "yosantanka_hiritsu_ijikanri_2": "0.000000", "yosantanka_hiritsu_ijikanri_3": "0.000000", "rakusatsu_ritsu": "0.950000", "reduc_shisetsu": "0.050000", "reduc_ijikanri_1": "0.050000", "reduc_ijikanri_2": "0.050000", "reduc_ijikanri_3": "0.020000", "pre_kyoukouka": true, "kisai_jutou": "0.75", "kisai_koufu": "0.299999999999999988897769753748434595763683319091796875", "hojo_ritsu": "0.299999999999999988897769753748434595763683319091796875", "zeimae_rieki": "0.000000", "SPC_keihi": "20.000000", "SPC_fee": "0.000000", "SPC_shihon": "0.000000", "SPC_yobihi": "0.000000", "SPC_hiyou_atsukai": 1, "houjinzei_ritsu": "0.0", "houjinjuminzei_kintou": "0.0", "hudousanshutokuzei_hyoujun": "0.0", "hudousanshutokuzei_ritsu": "0.0", "koteishisanzei_hyoujun": "0.0", "koteishisanzei_ritsu": "0.0", "tourokumenkyozei_hyoujun": "0.0", "tourokumenkyozei_ritsu": "0.0", "houjinjuminzei_ritsu_todouhuken": "0.0", "houjinjuminzei_ritsu_shikuchoson": "0.0"}}}
